add setting deadline to task on ajax
