+++ resources/views/layouts/app.blade.php
@@ -78,7 +78,8 @@
 <!-- Scripts -->
 <script src="{{ asset('js/app.js') }}"></script>
 <script src="{{ asset('js/ajax.js') }}"></script>
-<script src="{{ asset('js/project-ajax.js') }}"></script>
-<script src="{{ asset('js/task-ajax.js') }}"></script>
+<script src="{{ asset('js/project-crud-ajax.js') }}"></script>
+<script src="{{ asset('js/task-crud-ajax.js') }}"></script>
+<script src="{{ asset('js/task-extra-ajax.js') }}"></script>
 </body>
 </html>
