+++ resources/views/tasks/show-modal.blade.php
@@ -4,7 +4,6 @@
       <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
       <h4 class="modal-title text-center" id="show_task_modal_label">Task Information</h4>
     </div>
-    {{ Form::open() }}
     <div class="modal-body">
       <div class="container-fluid">
         <div class="row">
@@ -14,25 +13,23 @@
                 <table class="table table-striped table-hover">
                   <tr>
                     <th>Name</th>
-                    <td>$task->name</td>
+                    <td>{{ $task->name }}</td>
                   </tr>
                   <tr>
                     <th>Status</th>
                     @if ($task->status == 1)
-                      <td>Done</td>
+                    <td>Done</td>
                     @else
-                      <td>In process</td>
+                    <td>In process</td>
                     @endif
                   </tr>
                   <tr>
                     <th>Deadline</th>
-                    <td>
-                      {{ Form::text('deadline', '', ['class' => 'form-control deadline']) }}
-                      <script type="text/javascript">
-                        $(function() {
-                          $('.deadline').datepicker();
-                        });
-                      </script>
+                    <td class="deadline-container-js">
+                      <span class="span-deadline-js">{{ $task->deadline }}</span>
+                      {{ Form::open(['route' => ['tasks.deadline.edit', $task->id], 'class' => 'edit-task-deadline-form-js']) }}
+                        <button type="button" class="btn btn-default btn-xs btn-edit-task-deadline-task-js">Edit</button>
+                      {{ Form::close() }}
                     </td>
                   </tr>
                 </table>
@@ -45,6 +42,5 @@
     <div class="modal-footer">
       <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
     </div>
-    {{ Form::close() }}
   </div>
 </div>
